chore: Try Vite build for resolution, instead of `module-alias`
