+++ server/_api/v0/db/users/[uid]/[collectionId]/Params.d.ts
