+++ server/_api/v0/db/users/[uid]/[collectionId]/[documentId]/Params.d.ts
