+++ server/_api/v0/db/users/[uid]/attachments/[documentId]/Params.d.ts
