+++ server/_api/v0/logout/index.ts
@@ -1,0 +1,17 @@
+import { addJwtToBlacklist, jwtFromRequest, killSession } from "@/auth/jwt";
+import { apiHandler, dispatchRequests } from "@/helpers/apiHandler";
+import { respondSuccess } from "@/responses";
+
+export const POST = apiHandler("POST", async (req, res) => {
+	const token = jwtFromRequest(req, res);
+
+	killSession(req, res);
+
+	if (token !== null) {
+		await addJwtToBlacklist(token);
+	}
+
+	respondSuccess(res);
+});
+
+export default dispatchRequests({ POST });
