+++ server/auth/index.ts
@@ -2,15 +2,15 @@ import { asyncWrapper } from "@/asyncWrapper";
 import { Router } from "express";
 import { throttle } from "./throttle";
 
-import * as join from "@/api/v0/join";
-import * as login from "@/api/v0/login";
-import * as totpSecret from "@/api/v0/totp/secret";
-import * as totpValidate from "@/api/v0/totp/validate";
-import * as session from "@/api/v0/session";
-import * as logout from "@/api/v0/logout";
-import * as leave from "@/api/v0/leave";
-import * as updatepassword from "@/api/v0/updatepassword";
-import * as updateaccountid from "@/api/v0/updateaccountid";
+import * as join from "@/_api/v0/join";
+import * as login from "@/_api/v0/login";
+import * as totpSecret from "@/_api/v0/totp/secret";
+import * as totpValidate from "@/_api/v0/totp/validate";
+import * as session from "@/_api/v0/session";
+import * as logout from "@/_api/v0/logout";
+import * as leave from "@/_api/v0/leave";
+import * as updatepassword from "@/_api/v0/updatepassword";
+import * as updateaccountid from "@/_api/v0/updateaccountid";
 
 // TODO: Implement WebAuthn (and test passkey support!)
